moved to vanilla connection for single-spa
