+++ .vs/TinyCMS/xs/UserPrefs.xml
@@ -1,15 +1,8 @@
 ﻿<Properties StartupConfiguration="{79894BB9-CA52-4120-B1A9-21518C84C02C}|Default" NuGet.AddPackagesDialog.IncludePrerelease="True">
-  <MonoDevelop.Ide.Workbench ActiveDocument="TinyCMS.SocketServer/Extensions/NodeRequestExtensions.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="TinyCMS/Program.cs">
     <Files>
-      <File FileName="TinyCMS.SocketServer/MoveData.cs" Line="11" Column="41" />
-      <File FileName="TinyCMS.SocketServer/SocketRequest.cs" Line="56" Column="1" />
-      <File FileName="TinyCMS.SocketServer/SocketNodeServer.cs" Line="1" Column="1" />
-      <File FileName="TinyCMS.SocketServer/SocketConnectionHandler.cs" Line="69" Column="52" />
-      <File FileName="TinyCMS.SocketServer/Extensions/ContainerExtensions.cs" Line="15" Column="44" />
-      <File FileName="TinyCMS.SocketServer/Extensions/NodeRequestExtensions.cs" Line="69" Column="1" />
-      <File FileName="TinyCMS.Data/Extensions/NodeExtensions.cs" Line="89" Column="1" />
-      <File FileName="TinyCMS.Data/Builder/Container.cs" Line="127" Column="1" />
-      <File FileName="TinyCMS.Proxy/ProxyMiddleware.cs" Line="20" Column="1" />
+      <File FileName="TinyCMS/Startup.cs" Line="120" Column="47" />
+      <File FileName="TinyCMS/Program.cs" Line="9" Column="1" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.PinnedWatches />
