+++ TinyCMS/ReactClient/src/cms-link/AdminComponents/ObjectEditor.js
@@ -1,15 +1,16 @@
 import React from 'react';
 import PropertyEditor from './PropertyEditor';
 import NodeSelector from './NodeSelector';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { setEditorLink, onAuthenticationChanged } from '../connection';
+
+const closeIcon = (<svg aria-hidden="true" style={{ width: 20, height: 20 }} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 352 512"><path fill="currentColor" d="M242.72 256l100.07-100.07c12.28-12.28 12.28-32.19 0-44.48l-22.24-22.24c-12.28-12.28-32.19-12.28-44.48 0L176 189.28 75.93 89.21c-12.28-12.28-32.19-12.28-44.48 0L9.21 111.45c-12.28 12.28-12.28 32.19 0 44.48L109.28 256 9.21 356.07c-12.28 12.28-12.28 32.19 0 44.48l22.24 22.24c12.28 12.28 32.2 12.28 44.48 0L176 322.72l100.07 100.07c12.28 12.28 32.2 12.28 44.48 0l22.24-22.24c12.28-12.28 12.28-32.19 0-44.48L242.72 256z"></path></svg>);
 
 export const isReactNode = (dom) => {
     for (var key in dom) {
         if (key.startsWith("__reactInternalInstance$")) {
             var compInternals = dom[key];
             var owner = compInternals._debugOwner;
-            if (owner.memoizedProps && owner.memoizedProps.id) {
+            if (owner && owner.memoizedProps && owner.memoizedProps.id) {
                 return { node: owner.child.stateNode, owner, id: owner.memoizedProps.id };
             }
         }
@@ -43,11 +44,11 @@ export default class ObjectEditor extends React.Component {
             if (this._mounted)
                 this.forceUpdate();
         })
-        
+
 
         const fixButtons = (target) => {
             const div = document.createElement('div');
-            div.innerHTML = '<span>Edit</span>';
+            div.innerHTML = '<span><svg aria-hidden="true" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512" style="width:20px;height:16px"><path fill="currentColor" d="M290.74 93.24l128.02 128.02-277.99 277.99-114.14 12.6C11.35 513.54-1.56 500.62.14 485.34l12.7-114.22 277.9-277.88zm207.2-19.06l-60.11-60.11c-18.75-18.75-49.16-18.75-67.91 0l-56.55 56.55 128.02 128.02 56.55-56.55c18.75-18.76 18.75-49.16 0-67.91z" class=""></path></svg></span>';
             document.body.appendChild(div);
             target._editorNode = div;
             div.className = 'editor-button-overlay';
@@ -78,7 +79,6 @@ export default class ObjectEditor extends React.Component {
                 const currentTarget = findReactNode(e.target);
                 if (currentTarget && lastTarget.node !== currentTarget.node) {
                     fixButtons(currentTarget);
-                    //console.log(currentTarget);
                 }
             }
         });
@@ -119,7 +119,7 @@ export default class ObjectEditor extends React.Component {
                 onClick={() => {
                     this.setState({ isOpen: false });
                 }}>
-                <FontAwesomeIcon icon="times" />
+                {closeIcon}
             </span>
             <PropertyEditor ref={elm => { this.editor = elm }} match={{ params: { nodeId } }} />
             {/* <button type="button" className="btn btn-secondary" onClick={this.createNewSibling}>New sibling</button> */}
