+++ TinyCMS/ReactClient/src/scss/editor.scss
@@ -14,22 +14,20 @@ body {
     height:30px;
     background-color: lighten($topBarBg, 5);
     border-radius: 3px;
-    transition: all 0.3s ease-in-out;
+    transition: all 0.2s ease-in-out;
     opacity: 0;
     color:#fff;
-    padding:2px;
+    margin-left:-10px;
     cursor: pointer;
-    
     span {
-        display: block;
-        padding:2px 6px;
-        text-align: center;
-        border-radius: 3px;
-        border:solid 1px #fff;
+      display:block;
+      margin:5px;
     }
+    
   }
   
   .show {
+    margin-left:0px;
     opacity: 1;
   }
